fix uwp build openal-soft [sdl] fix uwp build sdl wip override final fix cmake SDL includes

+++ src/Plugins/SDLFileGroupPlugin/SDLFileInputStream.cpp
@@ -55,9 +55,9 @@ namespace Mengine
 
         if( 0 > size )
         {
-            LOGGER_ERROR( "invalid file '%s' size %d [error: %s]"
+            LOGGER_ERROR( "invalid file '%s' size %zu [error: %s]"
                 , fullPath
-                , size
+                , (size_t)size
                 , SDL_GetError()
             );
